Decided to leave data fetching inside uniProt archive
