--- src/rarediseasefinder/fetch/utils.py
@@ -1,15 +1,0 @@
-import requests
-
-def fetch_data(url):
-    try:
-        response = requests.get(url)
-        response.raise_for_status()
-        
-        return response.json()
-    
-    except requests.exceptions.HTTPError as http_err:
-        print(f"HTTP error occurred: {http_err}")
-    except Exception as err:
-        print(f"An error occurred: {err}")
-    
-    return None
